# backend/specs/003-survey-authoring/contracts/surveys.openapi.yaml
...
openapi: 3.1.0
info:
  title: Pitaco — Autoria de pesquisa
  version: 0.1.0
  description: >
    Superfície administrativa de autoria: criar a pesquisa, montar as perguntas, configurar
    o disparo e as regras, publicar congelando em versão imutável, controlar o que está no ar
    e versionar. Não entrega, não coleta e não agrega — isso é de outras features.
...
paths:
...
  /applications/{applicationId}/surveys/{surveyId}:
    get:
      tags: [Pesquisas]
      operationId: getSurvey
      summary: Consulta uma pesquisa
      description: >
        Devolve nome, estado derivado, aplicação dona, instante de criação e o conteúdo
        montado — o do rascunho quando existe, o da versão publicada caso contrário,
        indicado por `content.source`.
      parameters:
        - { $ref: '#/components/parameters/ApplicationId' }
        - { $ref: '#/components/parameters/SurveyId' }
      responses:
        '200':
          description: Pesquisa.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SurveyDetailResponse' }
        '404': { $ref: '#/components/responses/SurveyNotFound' }
    patch:
      tags: [Pesquisas]
      operationId: renameSurvey
      summary: Renomeia a pesquisa
      parameters:
        - { $ref: '#/components/parameters/ApplicationId' }
        - { $ref: '#/components/parameters/SurveyId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RenameSurveyRequest' }
      responses:
        '200':
          description: Nome alterado; o estado não muda.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SurveyResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/SurveyNotFound' }
    delete:
      tags: [Pesquisas]
      operationId: discardSurvey
      summary: Descarta uma pesquisa nunca publicada
      description: Publicada não se apaga, se encerra.
      parameters:
        - { $ref: '#/components/parameters/ApplicationId' }
        - { $ref: '#/components/parameters/SurveyId' }
      responses:
        '204': { description: Descartada, junto do conteúdo montado nela. }
        '404': { $ref: '#/components/responses/SurveyNotFound' }
        '422':
          description: Pesquisa já publicada (`survey.published_cannot_be_discarded`).
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/ApiError' } } }
...
  /applications/{applicationId}/surveys/{surveyId}/pause:
    post:
      tags: [Ciclo de vida]
      operationId: pauseSurvey
      summary: Pausa uma pesquisa ativa
      description: Imediato e reversível. A transição fica registrada com instante e motivo manual.
      parameters:
        - { $ref: '#/components/parameters/ApplicationId' }
        - { $ref: '#/components/parameters/SurveyId' }
      responses:
        '200':
          description: Pesquisa pausada.
          content: { application/json: { schema: { $ref: '#/components/schemas/SurveyResponse' } } }
        '404': { $ref: '#/components/responses/SurveyNotFound' }
        '422':
          description: >
            Pesquisa ainda em rascunho (`survey.not_published`) ou já encerrada
            (`survey.transition_not_allowed`).
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/ApiError' } } }
  /applications/{applicationId}/surveys/{surveyId}/resume:
    post:
      tags: [Ciclo de vida]
      operationId: resumeSurvey
      summary: Retoma uma pesquisa pausada
      description: Volta a ativa — ou a agendada, se a janela ainda não abriu.
      parameters:
        - { $ref: '#/components/parameters/ApplicationId' }
        - { $ref: '#/components/parameters/SurveyId' }
      responses:
        '200':
          description: Pesquisa retomada.
          content: { application/json: { schema: { $ref: '#/components/schemas/SurveyResponse' } } }
        '404': { $ref: '#/components/responses/SurveyNotFound' }
        '422':
          description: >
            Pesquisa em rascunho (`survey.not_published`), encerrada ou não pausada
            (`survey.transition_not_allowed`).
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/ApiError' } } }
  /applications/{applicationId}/surveys/{surveyId}/end:
    post:
      tags: [Ciclo de vida]
      operationId: endSurvey
      summary: Encerra uma pesquisa ativa ou pausada
      description: Definitivo. Nenhuma transição parte de encerrada.
      parameters:
        - { $ref: '#/components/parameters/ApplicationId' }
        - { $ref: '#/components/parameters/SurveyId' }
      responses:
        '200':
          description: Pesquisa encerrada.
          content: { application/json: { schema: { $ref: '#/components/schemas/SurveyResponse' } } }
        '404': { $ref: '#/components/responses/SurveyNotFound' }
        '422':
          description: >
            Pesquisa em rascunho (`survey.not_published`) ou já encerrada
            (`survey.transition_not_allowed`).
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/ApiError' } } }
  /applications/{applicationId}/surveys/{surveyId}/transitions:
    get:
      tags: [Ciclo de vida]
      operationId: listStateTransitions
      summary: Histórico de transições de estado
      description: >
        As transições comandadas somadas às derivadas da janela já passada, ordenadas por
        instante. Cada uma traz origem, destino, motivo e instante.
      parameters:
        - { $ref: '#/components/parameters/ApplicationId' }
        - { $ref: '#/components/parameters/SurveyId' }
      responses:
        '200':
          description: Histórico.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/StateTransitionResponse' }
        '404': { $ref: '#/components/responses/SurveyNotFound' }
...
components:
  parameters:
    ApplicationId:
      name: applicationId
      in: path
      required: true
      description: Identificador da aplicação dona.
      schema: { type: string, format: uuid }
    SurveyId:
      name: surveyId
      in: path
      required: true
      description: Identificador da pesquisa.
      schema: { type: string, format: uuid }
...
  responses:
    ValidationError:
      description: Requisição inválida, com o motivo de cada campo recusado.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ApiValidationError' }
    SurveyNotFound:
      description: >
        Pesquisa não encontrada (`survey.not_found`). Mesma resposta para identificador
        malformado, pesquisa inexistente e pesquisa de outra aplicação.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
  schemas:
...
    RenameSurveyRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
    SurveyState:
      type: string
      enum: [draft, scheduled, active, paused, ended]
      description: >
        Estado derivado do ciclo de vida e da janela no instante da leitura. `scheduled` vira
        `active` quando a janela abre, e `active` vira `ended` quando ela fecha; `paused` ignora
        a janela.
    SurveyResponse:
      type: object
      required: [id, applicationId, name, state, createdAt]
      properties:
        id: { type: string, format: uuid }
        applicationId: { type: string, format: uuid }
        name: { type: string }
        state: { $ref: '#/components/schemas/SurveyState' }
        publishedVersionNumber: { type: integer, nullable: true }
        draftVersionNumber: { type: integer, nullable: true }
        createdAt: { type: string, format: date-time }
    SurveyDetailResponse:
      allOf:
        - { $ref: '#/components/schemas/SurveyResponse' }
        - type: object
          properties:
            content:
              type: object
              nullable: true
              required: [source, versionNumber, questions]
              properties:
                source:
                  type: string
                  enum: [draft, published]
                  description: De onde veio o conteúdo devolvido.
                versionNumber: { type: integer }
                questions:
                  type: array
                  items: { $ref: '#/components/schemas/QuestionResponse' }
                trigger:
                  oneOf:
                    - { $ref: '#/components/schemas/TriggerResponse' }
                    - { type: 'null' }
...
    QuestionType:
      type: string
      enum: [single_choice, multiple_choice, rating, scale, nps, free_text]
    QuestionOption:
      type: object
      required: [label, value]
      properties:
        label: { type: string, maxLength: 200, example: Muito satisfeito }
        value: { type: string, maxLength: 120, example: very_satisfied }
...
    QuestionResponse:
      type: object
      required: [id, key, statement, type, position, required, options]
      properties:
        id: { type: string, format: uuid }
        key:
          type: string
          format: uuid
          description: Chave estável; atravessa versões e nunca muda.
        statement: { type: string }
        type: { $ref: '#/components/schemas/QuestionType' }
        position: { type: integer, minimum: 1 }
        required: { type: boolean }
        options:
          type: array
          items: { $ref: '#/components/schemas/QuestionOption' }
        range:
          type: object
          nullable: true
          properties:
            min: { type: integer }
            max: { type: integer }
...
    TriggerRequest:
      type: object
      required: [eventName, windowStart, samplingRate]
      properties:
        eventName:
          type: string
          pattern: '^[a-z][a-z0-9_.]{1,79}$'
          example: checkout.completed
        windowStart: { type: string, format: date-time }
        windowEnd:
          type: string
          format: date-time
          nullable: true
          description: Ausente significa tempo indeterminado.
        samplingRate: { type: number, minimum: 0, maximum: 1, example: 0.25 }
    TriggerResponse:
      allOf:
        - { $ref: '#/components/schemas/TriggerRequest' }
        - type: object
          properties:
            rules:
              type: array
              items: { $ref: '#/components/schemas/SegmentationRuleResponse' }
    RuleOperation:
      type: string
      enum: [equals, not_equals, present, absent]
    SegmentationRuleRequest:
      type: object
      required: [attribute, operation]
      description: >
        `value` é exigido em `equals` e `not_equals`, e recusado em `present` e `absent`.
      properties:
        attribute: { type: string, minLength: 1, maxLength: 80, example: plan }
        operation: { $ref: '#/components/schemas/RuleOperation' }
        value: { type: string, maxLength: 200, nullable: true, example: pro }
    SegmentationRuleResponse:
      allOf:
        - { $ref: '#/components/schemas/SegmentationRuleRequest' }
        - type: object
          required: [id]
          properties:
            id: { type: string, format: uuid }
...
    StateTransitionResponse:
      type: object
      required: [from, to, reason, occurredAt]
      properties:
        from: { $ref: '#/components/schemas/SurveyState' }
        to: { $ref: '#/components/schemas/SurveyState' }
        reason:
          type: string
          enum: [publication, manual_pause, manual_resume, manual_end, window_opened, window_closed]
          description: >
            As quatro primeiras são comandadas e ficam registradas; as duas últimas são derivadas
            da janela já passada e calculadas na leitura.
        actor:
          type: string
          nullable: true
          description: Reservado; sempre ausente enquanto não houver autenticação.
        occurredAt: { type: string, format: date-time }
...
    ApiError:
      type: object
      required: [title, status, detail, instance, code]
      description: RFC 9457, gerado exclusivamente pelo ApiExceptionHandler.
      properties:
        type: { type: string, example: about:blank }
        title: { type: string, example: Unprocessable Content }
        status: { type: integer, example: 422 }
        detail: { type: string }
        instance: { type: string, example: '/api/applications/{applicationId}/surveys/{surveyId}/publication' }
        code: { type: string, example: survey.not_publishable }
        traceId: { type: string, nullable: true }
    ApiValidationError:
      allOf:
        - { $ref: '#/components/schemas/ApiError' }
        - type: object
          required: [errors]
          properties:
            errors:
              type: object
              additionalProperties: { type: string }
              example: { statement: 'Enunciado é obrigatório' }